+++ types/next.d.ts
@@ -1966,7 +1966,8 @@ export namespace xdr {
       | 'scvAddress'
       | 'scvContractInstance'
       | 'scvLedgerKeyContractInstance'
-      | 'scvLedgerKeyNonce';
+      | 'scvLedgerKeyNonce'
+      | 'scvExecutableTag';
 
     readonly value:
       | 0
@@ -1990,7 +1991,8 @@ export namespace xdr {
       | 18
       | 19
       | 20
-      | 21;
+      | 21
+      | 22;
 
     static scvBool(): ScValType;
 
@@ -2035,6 +2037,8 @@ export namespace xdr {
     static scvLedgerKeyContractInstance(): ScValType;
 
     static scvLedgerKeyNonce(): ScValType;
+
+    static scvExecutableTag(): ScValType;
   }
 
   class ScErrorType {
@@ -2110,13 +2114,18 @@ export namespace xdr {
   }
 
   class ContractExecutableType {
-    readonly name: 'contractExecutableWasm' | 'contractExecutableStellarAsset';
+    readonly name:
+      | 'contractExecutableWasm'
+      | 'contractExecutableStellarAsset'
+      | 'contractExecutableExternalRef';
 
-    readonly value: 0 | 1;
+    readonly value: 0 | 1 | 2;
 
     static contractExecutableWasm(): ContractExecutableType;
 
     static contractExecutableStellarAsset(): ContractExecutableType;
+
+    static contractExecutableExternalRef(): ContractExecutableType;
   }
 
   class ScAddressType {
@@ -15452,11 +15461,19 @@ export namespace xdr {
 
     wasmHash(value?: Buffer): Buffer;
 
+    externalRef(
+      value?: ContractExecutableExternalRef,
+    ): ContractExecutableExternalRef;
+
     static contractExecutableWasm(value: Buffer): ContractExecutable;
 
     static contractExecutableStellarAsset(): ContractExecutable;
 
-    value(): Buffer | void;
+    static contractExecutableExternalRef(
+      value: ContractExecutableExternalRef,
+    ): ContractExecutable;
+
+    value(): Buffer | ContractExecutableExternalRef | void;
 
     toXDR(format?: 'raw'): Buffer;
 
@@ -15473,6 +15490,43 @@ export namespace xdr {
     static fromXDR(input: Buffer, format?: 'raw'): ContractExecutable;
 
     static fromXDR(input: string, format: 'hex' | 'base64'): ContractExecutable;
+
+    static validateXDR(input: Buffer, format?: 'raw'): boolean;
+
+    static validateXDR(input: string, format: 'hex' | 'base64'): boolean;
+  }
+
+  class ContractExecutableExternalRef {
+    constructor(attributes: {
+      executableOwner: ScAddress;
+      tag: string | Buffer;
+    });
+
+    executableOwner(value?: ScAddress): ScAddress;
+
+    tag(value?: string | Buffer): string | Buffer;
+
+    toXDR(format?: 'raw'): Buffer;
+
+    toXDR(format: 'hex' | 'base64'): string;
+
+    static read(io: Buffer): ContractExecutableExternalRef;
+
+    static write(value: ContractExecutableExternalRef, io: Buffer): void;
+
+    static isValid(value: ContractExecutableExternalRef): boolean;
+
+    static toXDR(value: ContractExecutableExternalRef): Buffer;
+
+    static fromXDR(
+      input: Buffer,
+      format?: 'raw',
+    ): ContractExecutableExternalRef;
+
+    static fromXDR(
+      input: string,
+      format: 'hex' | 'base64',
+    ): ContractExecutableExternalRef;
 
     static validateXDR(input: Buffer, format?: 'raw'): boolean;
 
@@ -15573,6 +15627,8 @@ export namespace xdr {
 
     nonceKey(value?: ScNonceKey): ScNonceKey;
 
+    executableTag(value?: string | Buffer): string | Buffer;
+
     static scvBool(value: boolean): ScVal;
 
     static scvVoid(): ScVal;
@@ -15616,6 +15672,8 @@ export namespace xdr {
     static scvLedgerKeyContractInstance(): ScVal;
 
     static scvLedgerKeyNonce(value: ScNonceKey): ScVal;
+
+    static scvExecutableTag(value: string | Buffer): ScVal;
 
     value():
       | boolean